refactor error handling and move it to thunk
